Fix: total set solutions and optimised for n_used.
The values of `indexes[i] forall i if is_unused(i)` were being constrained
to the value of `i` when these values are often placed outside the range
of feasible `indexes[i]` values.

Also added a term in the objective function to optimise for brevity of
solution with a much smaller weighting than the distance to the target.